Merge v3.01.00-RC into `develop`

+++ server/graphql/core/src/loader/item_stock_on_hand.rs
@@ -26,10 +26,10 @@ impl ItemsStockOnHandLoaderInput {
 /// `available` excludes held/reserved packs; `total` is all packs. Exposing both
 /// lets `ItemNode.stockOnHand` be served from this loader instead of the much more
 /// expensive item-stats path (which runs the AMC backend plugin).
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+#[derive(Clone, Copy, Debug, PartialEq, Default)]
 pub struct ItemStockOnHand {
-    pub available_stock_on_hand: u32,
-    pub total_stock_on_hand: u32,
+    pub available_stock_on_hand: f64,
+    pub total_stock_on_hand: f64,
 }
 
 impl Loader<ItemsStockOnHandLoaderInput> for ItemsStockOnHandLoader {
@@ -64,8 +64,8 @@ impl Loader<ItemsStockOnHandLoaderInput> for ItemsStockOnHandLoader {
                         &stock_on_hand.item_id,
                     ),
                     ItemStockOnHand {
-                        available_stock_on_hand: stock_on_hand.available_stock_on_hand as u32,
-                        total_stock_on_hand: stock_on_hand.total_stock_on_hand as u32,
+                        available_stock_on_hand: stock_on_hand.available_stock_on_hand,
+                        total_stock_on_hand: stock_on_hand.total_stock_on_hand,
                     },
                 )
             })
